Rando release

+++ OpenKh.Tools.ModsManager/ViewModels/MainViewModel.cs
@@ -41,6 +41,7 @@ namespace OpenKh.Tools.ModsManager.ViewModels
         private bool _panaceaInstalled;
         private bool _panaceaConsoleEnabled;
         private bool _panaceaDebugLogEnabled;
+        private bool _panaceaSoundDebugEnabled;
         private bool _panaceaCacheEnabled;
         private bool _panaceaQuickMenuEnabled;
         private bool _devView;
@@ -54,7 +55,7 @@ namespace OpenKh.Tools.ModsManager.ViewModels
             "Recom",
             "kh3d"
         };
-        private int _wizardVersionNumber = 4;
+        private int _wizardVersionNumber = 5;
         private string[] executable = new string[]
         {
             "KINGDOM HEARTS II FINAL MIX.exe",
@@ -142,9 +143,23 @@ namespace OpenKh.Tools.ModsManager.ViewModels
             {
                 _panaceaDebugLogEnabled = value;
                 ConfigurationService.DebugLog = _panaceaDebugLogEnabled;
+                if (_panaceaSoundDebugEnabled)
+                    PanaceaSoundDebugEnabled = false;
                 OnPropertyChanged(nameof(PanaceaDebugLogEnabled));
                 UpdatePanaceaSettings();
             }
+        }
+        public bool PanaceaSoundDebugEnabled
+        {
+            get => _panaceaSoundDebugEnabled;
+            set
+            {
+                _panaceaSoundDebugEnabled = value;
+                ConfigurationService.SoundDebug = _panaceaSoundDebugEnabled;
+                OnPropertyChanged(nameof(PanaceaSoundDebugEnabled));
+                UpdatePanaceaSettings();
+            }
+
         }
         public bool PanaceaCacheEnabled
         {
@@ -299,6 +314,7 @@ namespace OpenKh.Tools.ModsManager.ViewModels
                 DevView = ConfigurationService.DevView;
                 _panaceaConsoleEnabled = ConfigurationService.ShowConsole;
                 _panaceaDebugLogEnabled = ConfigurationService.DebugLog;
+                _panaceaSoundDebugEnabled = ConfigurationService.SoundDebug;
                 _panaceaCacheEnabled = ConfigurationService.EnableCache;
                 _panaceaQuickMenuEnabled = ConfigurationService.QuickMenu;
             }
@@ -705,7 +721,7 @@ namespace OpenKh.Tools.ModsManager.ViewModels
                             return Task.CompletedTask;
                         }
                     }
-                    if (ConfigurationService.PCVersion == "Steam" && !(_launchGame == "kh3d"))
+                    if (ConfigurationService.PCVersion == "Steam" && !(_launchGame == "kh3d") && ConfigurationService.SteamAPITrick1525 == false)
                     {
                         if (ConfigurationService.PcReleaseLocation != null)
                         {
@@ -741,7 +757,7 @@ namespace OpenKh.Tools.ModsManager.ViewModels
                             return Task.CompletedTask;
                         }
                     }
-                    else if (ConfigurationService.PCVersion == "Steam" && _launchGame == "kh3d")
+                    else if (ConfigurationService.PCVersion == "Steam" && _launchGame == "kh3d" && ConfigurationService.SteamAPITrick28 == false)
                     {
                         if (ConfigurationService.PcReleaseLocationKH3D != null)
                         {
@@ -941,8 +957,11 @@ namespace OpenKh.Tools.ModsManager.ViewModels
                     {
                         var sourceFile = Path.Combine(ConfigurationService.GameModPath, _launchGame, entry.Key);
                         var destFile = Path.Combine(patchStagingDir, entry.Value);
-                        Directory.CreateDirectory(Path.GetDirectoryName(destFile));
-                        File.Move(sourceFile, destFile);
+                        if (File.Exists(sourceFile))
+                        {
+                            Directory.CreateDirectory(Path.GetDirectoryName(destFile));
+                            File.Move(sourceFile, destFile);
+                        }
                     }
 
                     foreach (var directory in Directory.GetDirectories(Path.Combine(ConfigurationService.GameModPath, _launchGame)))
@@ -1271,7 +1290,8 @@ namespace OpenKh.Tools.ModsManager.ViewModels
                         }
                     }
                     textToWrite += $"\r\nshow_console={_panaceaConsoleEnabled}\r\n" +
-                        $"debug_log={_panaceaDebugLogEnabled}\r\nenable_cache={_panaceaCacheEnabled}\r\nquick_menu={_panaceaQuickMenuEnabled}";
+                        $"debug_log={_panaceaDebugLogEnabled}\r\nsound_debug={_panaceaSoundDebugEnabled}\r\n" +
+                        $"enable_cache={_panaceaCacheEnabled}\r\nquick_menu={_panaceaQuickMenuEnabled}";
                     File.WriteAllText(panaceaSettings, textToWrite);
                 }
                 else if (ConfigurationService.PcReleaseLocationKH3D != null)
@@ -1288,7 +1308,8 @@ namespace OpenKh.Tools.ModsManager.ViewModels
                         }
                     }
                     textToWrite += $"\r\nshow_console={_panaceaConsoleEnabled}\r\n" +
-                        $"debug_log={_panaceaDebugLogEnabled}\r\nenable_cache={_panaceaCacheEnabled}\r\nquick_menu={_panaceaQuickMenuEnabled}";
+                        $"debug_log={_panaceaDebugLogEnabled}\r\nsound_debug={_panaceaSoundDebugEnabled}\r\n" +
+                        $"enable_cache={_panaceaCacheEnabled}\r\nquick_menu={_panaceaQuickMenuEnabled}";
                     File.WriteAllText(panaceaSettings, textToWrite);
                 }
             }
